generateSound/requirements.txt
appnope==0.1.0
audioread==2.1.5
bleach==1.5.0
bokeh==0.12.13
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
future==0.16.0
futures==3.1.1
google-cloud-storage==1.6.0
html5lib==0.9999999
intervaltree==2.1.0
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
librosa==0.5.1
llvmlite==0.21.0
magenta==0.3.5
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.1
mido==1.2.6
mir-eval==0.4
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numba==0.36.2
numpy==1.14.0
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pretty-midi==0.2.8
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-rtmidi==1.1.0
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
resampy==0.2.0
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sortedcontainers==1.5.9
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.1.0
